Autotoolize

+++ maxlocal.c
@@ -46,7 +46,7 @@ main (int argc, char const* argv[])
                                       &max_local_size,
                                       NULL));
 
-    printf ("Local mem size: %ld bytes\n", max_local_size);
+    printf ("Local mem size: %d bytes\n", (unsigned) max_local_size);
 
     source = make_kernel (max_local_size);
     program = ocl_create_program_from_source (ocl, source, NULL, &errcode);
